directory organization, server renamed
