+++ instruments/digitalout.py
