+++ tests/cmd_tests.py
